[Feat] 가끔 유닛이 안 죽던 버그 해결
문제 : 한 프레임에 여러번 상태전환이일어나며 꼬인 것
