+++ Assets/Workspace/HSD/Scripts/Unit/FSM/StateMachine.cs
@@ -7,6 +7,7 @@ public class StateMachine
 {
     public BaseState _currentState;
     private BaseState _nextState;
+    private bool _isTransitioning;
 
     public void Update()
     {
@@ -21,6 +22,8 @@ public class StateMachine
 
     public void ChangeState(BaseState newState)
     {
+        if (_isTransitioning) return;
+        _isTransitioning = true;
         _nextState = newState;
     }
 
@@ -29,5 +32,6 @@ public class StateMachine
         _currentState?.Exit();
         _currentState = newState;
         _currentState?.Enter();
+        _isTransitioning = false;
     }
 }
