+++ Assets/Workspace/HSD/Scripts/Unit/UnitFXController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 83f62c7108c64294fb57bb62f6e0cf56
+guid: f8b6dbb8cfa5ceb44a9cc436083b9a96
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
